New benchmarking arch

See merge request dev-hood/elpida/elpida!7

+++ src/Elpida/Utilities/MemoryFile.cpp
@@ -27,25 +27,22 @@
 #include "Elpida/Utilities/MemoryFile.hpp"
 
 #include <fstream>
+#include "Elpida/Config.hpp"
 #include "Elpida/ElpidaException.hpp"
+#include "Elpida/Utilities/NumaMemory.hpp"
+#include "Elpida/Topology/SystemTopology.hpp"
 
 namespace Elpida
 {
 
-	MemoryFile::MemoryFile()
-		: _data(nullptr), _size(0), _deleteData(true)
+	MemoryFile::MemoryFile(unsigned int processorAffinity)
+		: _data(nullptr), _allocatedMemory(nullptr), _processorAffinity(processorAffinity), _size(0), _deleteData(true)
 	{
 
-	}
-
-	MemoryFile::MemoryFile(std::size_t size)
-		: _size(size), _deleteData(true)
-	{
-		_data = new Data[_size];
 	}
 
 	MemoryFile::MemoryFile(void* data, std::size_t size)
-		: _data((DataPtr)data), _size(size), _deleteData(false)
+		: _data((pData)data), _allocatedMemory(nullptr), _processorAffinity(-1), _size(size), _deleteData(false)
 	{
 
 	}
@@ -57,34 +54,44 @@ namespace Elpida
 
 	void MemoryFile::load(const std::string& path)
 	{
-		destroyData();
-		_size = 0;
-		std::ifstream file(path, std::ifstream::binary);
-		try
-		{
-			file.exceptions(std::ios::failbit);
-			file.seekg(0, std::ifstream::end);
-			_size = file.tellg();
-			file.seekg(0, std::ifstream::beg);
-
-			_data = new Data[_size];
-			file.read((char*)_data, _size);
-		}
-		catch (...)
+		if (_deleteData)
 		{
 			destroyData();
-			if (file.is_open())
+			_size = 0;
+			std::ifstream file(path, std::ifstream::binary);
+			try
 			{
-				file.close();
+				file.exceptions(std::ios::failbit);
+				file.seekg(0, std::ifstream::end);
+				_size = file.tellg();
+				file.seekg(0, std::ifstream::beg);
+
+				allocateData();
+				file.read((char*)_data, _size);
 			}
-			throw;
+			catch (...)
+			{
+				destroyData();
+				if (file.is_open())
+				{
+					file.close();
+				}
+				throw;
+			}
 		}
+		else
+		{
+			throw ElpidaException(FUNCTION_NAME, "Attempted to load a file when it was constructed with existing data");
+		}
+
 	}
+
 	void MemoryFile::destroyData()
 	{
 		if (_deleteData)
 		{
-			delete[] _data;
+			delete _allocatedMemory;
+			_allocatedMemory = nullptr;
 		}
 	}
 
@@ -106,6 +113,13 @@ namespace Elpida
 			}
 			throw;
 		}
+	}
+
+	void MemoryFile::allocateData()
+	{
+		_allocatedMemory = new NumaMemory(_size, SystemTopology::getNumaNodeOfProcessor(_processorAffinity));
+		_allocatedMemory->allocate();
+		_data = static_cast<pData>(_allocatedMemory->getPointer());
 	}
 
 } /* namespace Elpida */
